feat: добавлена логика запрета удаления комнаты при активных бронях

+++ RoomsBooking.Application/UseCases/Rooms/Handlers/DeleteRoomByIdCommandHandler.cs
@@ -2,17 +2,24 @@
 using Microsoft.EntityFrameworkCore;
 using RoomsBooking.Application.Interfaces;
 using RoomsBooking.Application.UseCases.Rooms.Commands;
+using RoomsBooking.Domain.Exceptions.Room;
 
 namespace RoomsBooking.Application.UseCases.Rooms.Handlers;
 
-// TODO: Когда будет готово бронирование комнат, надо явно запрещать удаление если есть активные брони или делать мягкое удаление
 public class DeleteRoomByIdCommandHandler(
     IAppDbContext dbContext)
     : IRequestHandler<DeleteRoomByIdCommand, Unit>
 {
     public async Task<Unit> Handle(DeleteRoomByIdCommand request, CancellationToken cancellationToken)
     {
+        var hasActiveBookings = await dbContext.Bookings
+            .AnyAsync(b => b.RoomId == request.Id && b.EndTime > DateTimeOffset.UtcNow, cancellationToken);
+
+        if (hasActiveBookings)
+            throw new RoomHasActiveBookingsException(request.Id);
+
         await dbContext.Rooms.Where(r => r.Id == request.Id).ExecuteDeleteAsync(cancellationToken);
+
         return Unit.Value;
     }
 }
